app/requirements.txt
Django
requests
bs4
redis
python-socketio
eventlet
redis
celery
python-dotenv
Pillow
djangorestframework
drf-yasg